requirements.txt
paramiko
llama-cpp-python
pandas
psycopg2-binary
